pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Production-grade RAG pipeline with hybrid retrieval, citation verification, and confidence scoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.11.0",
    "sentence-transformers>=3.0.0",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
    "rank-bm25>=0.2.2",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "streamlit>=1.36.0",
    "pyyaml>=6.0",
    "pydantic>=2.7.0",
    "jinja2>=3.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.6",
    "PyPDF2>=3.0.0",
    "numpy>=1.26.0",
]

[project.scripts]
rag = "rag.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
